Drop full compatibility with scipy in favor of more sensible distances
In several of the scipy distance functions, the vectors were already
assumed to be coded as 1 and 0 to indicate presence/absence values. To
generate an indicator for values indicating absence, the autors of scipy
subtracted the vector from 1.  For vectors coded as 0 and 1, this had the
effect of changing the 1's to 0's and the 0's to 1's, as intended.
However, for vectors not coded as 0 and 1, this transformation resulted
in nonsensical distances.

Here, we move away from complete compatibility with scipy, and adopt
the more straightforward method of first converting all values to
presence/absence right away.  When an indicator for absence is needed,
we use a boolean function, instead of subtraction.  This results in
distances that behave well and are consistent with the rest of our library.