Remove checks in alpha diversity functions
Previously, we had checked for negative vectors and for
vectors that summed to zero. These generated a warning and
a return value of NA.

We have removed these checks to increase the simplicity of
the alpha diversity functions. Without the checks, most
functions return NaN on vectors of zeros, which we document
in the tests.